Working Crank & Arcade button logic!
Re-thought myself through the crank sampling algorithm, as well as the 
main button and crank logic.

Was running into a bug where just the arcade button was working good, 
just the crank would work good, but if I started cranking when the 
button was active or vice versa, I'd get a second set of sounds that 
never go away.  The part around line 550 was the fix to that.