Added a soundKill() function to the strings
Per request from John.  Releasing the toggle while not cranking now 
sends a ControlChange 123 (All notes off) to all strings, acting as a 
sort of kill button.

Hopefully this helps in odd situations where the synth is stuck on 
(unclear if that's the synth stuck on, or the digigurdy somehow).