2.1.9 - part 1 - Tsunami/wavTrigger License file
Not sure what the license is supposed to be exactly since I can't find 
on on the repo, but I assume they are for open use since they're 
publicly available and it's from the creator of the board we're 
purchasing/using.

I wanted to put this here to be clear that the wavTrigger and Tsunai 
code is not my creation.